chore: factor out commandid from the server lib
Also, move command registry command logic to main_service
where it should reside.
