+++ src/server/command_registry.h
@@ -10,7 +10,7 @@
 #include <functional>
 
 #include "base/function2.hpp"
-#include "server/common.h"
+#include "facade/command_id.h"
 
 namespace dfly {
 
@@ -53,8 +53,11 @@ static_assert(!IsEvalKind(""));
 
 };  // namespace CO
 
-class CommandId {
+class CommandId : public facade::CommandId {
  public:
+  CommandId(const char* name, uint32_t mask, int8_t arity, int8_t first_key, int8_t last_key,
+            int8_t step);
+
   using Handler =
       fu2::function_base<true /*owns*/, true /*copyable*/, fu2::capacity_default,
                          false /* non-throwing*/, false /* strong exceptions guarantees*/,
@@ -63,52 +66,8 @@ class CommandId {
   using ArgValidator = fu2::function_base<true, true, fu2::capacity_default, false, false,
                                           bool(CmdArgList, ConnectionContext*) const>;
 
-  /**
-   * @brief Construct a new Command Id object
-   *
-   * When creating a new command use the https://github.com/redis/redis/tree/unstable/src/commands
-   * files to find the right arguments.
-   *
-   * @param name
-   * @param mask
-   * @param arity -     positive if command has fixed number of required arguments
-   *                    negative if command has minimum number of required arguments, but may have
-   * more.
-   * @param first_key - position of first key in argument list
-   * @param last_key  - position of last key in argument list,
-   *                    -1 means the last key index is (arg_length - 1), -2 means that the last key
-   * index is (arg_length - 2).
-   * @param step -      step count for locating repeating keys
-   */
-  CommandId(const char* name, uint32_t mask, int8_t arity, int8_t first_key, int8_t last_key,
-            int8_t step);
-
-  std::string_view name() const {
-    return name_;
-  }
-
-  int arity() const {
-    return arity_;
-  }
-
-  uint32_t opt_mask() const {
-    return opt_mask_;
-  }
-
-  int8_t first_key_pos() const {
-    return first_key_;
-  }
-
-  int8_t last_key_pos() const {
-    return last_key_;
-  }
-
   bool is_multi_key() const {
     return (last_key_ != first_key_) || (opt_mask_ & CO::VARIADIC_KEYS);
-  }
-
-  int8_t key_arg_step() const {
-    return step_key_;
   }
 
   CommandId& SetHandler(Handler f) {
@@ -126,25 +85,16 @@ class CommandId {
     handler_(std::move(args), cntx);
   }
 
-  bool IsTransactional() const;
-
   // Returns true if validation succeeded.
   bool Validate(CmdArgList args, ConnectionContext* cntx) const {
     return !validator_ || validator_(std::move(args), cntx);
   }
 
+  bool IsTransactional() const;
+
   static const char* OptName(CO::CommandOpt fl);
-  static uint32_t OptCount(uint32_t mask);
 
  private:
-  std::string_view name_;
-
-  uint32_t opt_mask_;
-  int8_t arity_;
-  int8_t first_key_;
-  int8_t last_key_;
-  int8_t step_key_;
-
   Handler handler_;
   ArgValidator validator_;
 };
@@ -153,7 +103,7 @@ class CommandRegistry {
   absl::flat_hash_map<std::string_view, CommandId> cmd_map_;
 
  public:
-  CommandRegistry();
+  CommandRegistry() = default;
 
   CommandRegistry& operator<<(CommandId cmd);
 
@@ -174,10 +124,6 @@ class CommandRegistry {
       cb(k_v.first, k_v.second);
     }
   }
-
- private:
-  // Implements COMMAND functionality.
-  void Command(CmdArgList args, ConnectionContext* cntx);
 };
 
 }  // namespace dfly
